rename items.py to inventory.py
